nxutilからEdgeTypeをファイル分割
