+++ knowde/primitive/__core__/nxutil/__init__.py
@@ -1,15 +1,13 @@
 """networkx関連."""
 from __future__ import annotations
 
-from enum import Enum, auto
-from functools import cache, cached_property
+from functools import cache
 from pprint import pp
-from typing import Any, Callable, Hashable, Iterator
+from typing import Any, Hashable
 
 import networkx as nx
 
-from knowde.primitive.__core__.nxutil.errors import MultiEdgesError
-
+from .edge_type import Direction, EdgeType  # noqa: F401
 from .types import Accessor, Edges
 
 
@@ -34,10 +32,6 @@ def to_nested(g: nx.DiGraph, start: Hashable, f: Accessor) -> dict:
     return _f(start)
 
 
-# def to_list(g: nx.DiGraph, start: Hashable, f: Accessor) -> list[Hashable]:
-#     """有向グラフから関連をたどって."""
-
-
 def to_nodes(g: nx.DiGraph, start: Hashable, f: Accessor) -> list[Hashable]:
     """有向グラフを辿ってノードを取得."""
     s = []
@@ -49,30 +43,6 @@ def to_nodes(g: nx.DiGraph, start: Hashable, f: Accessor) -> list[Hashable]:
 
     _f(start)
     return s
-
-
-def succ_attr(attr_name: str, value: Any) -> Accessor:  # noqa: ANN401
-    """次を関係の属性から辿る."""
-
-    def _f(g: nx.DiGraph, start: Hashable) -> Iterator[Hashable]:
-        for _, succ, d in g.out_edges(start, data=True):
-            if any(d) and d[attr_name] == value:
-                _f(g, succ)
-                yield succ
-
-    return _f
-
-
-def pred_attr(attr_name: str, value: Any) -> Accessor:  # noqa: ANN401
-    """前を関係の属性から辿る."""
-
-    def _f(g: nx.DiGraph, start: Hashable) -> Iterator[Hashable]:
-        for pred, _, d in g.in_edges(start, data=True):
-            if any(d) and d[attr_name] == value:
-                _f(g, pred)
-                yield pred
-
-    return _f
 
 
 def filter_edge_attr(g: nx.DiGraph, name: str, value: Any) -> nx.DiGraph:  # noqa: ANN401
@@ -113,111 +83,6 @@ def axiom_paths(g: nx.DiGraph, tgt: Hashable, t: Any) -> list[list[Hashable]]:  
     return _to_paths(sub, pairs)
 
 
-def nxconvert(g: nx.DiGraph, convert: Callable[[Hashable], Hashable]) -> nx.DiGraph:
-    """ネットワークのノード型を変換."""
-    new = nx.DiGraph()
-    new.add_nodes_from([convert(n) for n in g.nodes])
-    new.add_edges_from([(convert(u), convert(v), d) for u, v, d in g.edges(data=True)])
-    return new
-
-
-class Direction(Enum):
-    """方向."""
-
-    FORWARD = auto()
-    BACKWARD = auto()
-    BOTH = auto()
-
-
-class EdgeType(Enum):
-    """グラフ関係の種類."""
-
-    # 文章構成
-    HEAD = auto()  # 見出しを配下にする
-    SIBLING = auto()  # 兄弟 同階層 並列
-    BELOW = auto()  # 配下 階層が下がる 直列
-
-    # 意味的関係
-    DEF = auto()  # term -> 文
-    RESOLVED = auto()  # 用語解決関係 文 -> 文
-    TO = auto()  # 依存
-    EXAMPLE = auto()  # 具体
-
-    # 付加的
-    WHEN = auto()
-    WHERE = auto()
-    NUM = auto()
-    BY = auto()
-    REF = auto()
-
-    # both
-    ANTI = auto()  # 反対
-    SIMILAR = auto()  # 類似
-
-    @property
-    def forward(self) -> tuple[EdgeType, Direction]:
-        """正順."""
-        return self, Direction.FORWARD
-
-    @property
-    def backward(self) -> tuple[EdgeType, Direction]:
-        """逆順."""
-        return self, Direction.BACKWARD
-
-    @property
-    def both(self) -> tuple[EdgeType, Direction]:
-        """両順."""
-        return self, Direction.BOTH
-
-    def add_edge(self, g: nx.DiGraph, pre: Hashable, suc: Hashable) -> None:
-        """エッジ追加."""
-        g.add_edge(pre, suc, type=self)
-
-    def add_path(
-        self,
-        g: nx.DiGraph,
-        *ns: Hashable,
-        cvt: Callable[[Hashable], Hashable] = lambda x: x,
-    ) -> list[Hashable]:
-        """連続追加."""
-        ls = [cvt(n) for n in ns]
-        nx.add_path(g, ls, type=self)
-        return ls
-
-    @cached_property
-    def succ(self) -> Accessor:
-        """エッジを辿って次を取得."""
-        return succ_attr("type", self)
-
-    @cached_property
-    def pred(self) -> Accessor:
-        """エッジを遡って前を取得."""
-        return pred_attr("type", self)
-
-    def get_succ_or_none(self, g: nx.DiGraph, n: Hashable) -> None | Hashable:
-        """1つの先を返す."""
-        return _get_one_or_none(list(self.succ(g, n)), self, n)
-
-    def get_pred_or_none(self, g: nx.DiGraph, n: Hashable) -> None | Hashable:
-        """1つの前を返す."""
-        return _get_one_or_none(list(self.pred(g, n)), self, n)
-
-
-def _get_one_or_none(ls: list[Hashable], t: EdgeType, src: Hashable) -> None | Hashable:
-    """1つまたはなしを取得."""
-    match len(ls):
-        case 0:
-            return None
-        case 1:
-            return ls[0]
-        case _:
-            msg = (
-                f"'{src}'から複数の関係がヒットしました. 1つだけに修正してね: {t} \n\t"
-                + "\n\t".join(map(str, ls))
-            )
-            raise MultiEdgesError(msg)
-
-
 def replace_node(g: nx.DiGraph, old: Hashable, new: Hashable) -> None:
     """エッジを保ってノードを置換."""
     g.add_node(new)
@@ -227,7 +92,3 @@ def replace_node(g: nx.DiGraph, old: Hashable, new: Hashable) -> None:
     for _, succ, data in g.out_edges(old, data=True):
         g.add_edge(new, succ, **data)
     g.remove_node(old)
-
-
-# def add_siblings(g: nx.DiGraph, old: Hashable, new: Hashable) -> None:
-#     """Old BELOW を new SIBLINGに追加する."""
